+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Dev
 
 - Added multiple containers support(instead of scopes)
+- Added Resolver class
+- Container is refactored to use Resolver class
 
 ## 1.0.1